requirements.txt
streamlit
google-generativeai
xhtml2pdf
python-docx
beautifulsoup4
pillow